Add Water Heater
This commit adds a first version of the water heater support for the
K30.

Though I think it would be even better if we split up the creation of
the entity based on the references we have saved in the bhc.

Unfortunatly I'm only able to test for 1 dhw, as thats the setup I have.